Refactor: harmonic.c and harmonic.h
